GraphQL TypeScript Setup: graphql/schema
