+++ src/index.ts
@@ -1,43 +1,11 @@
-// import { ApolloServer } from "apollo-server";
-
 import { ApolloServer, gql } from "apollo-server";
 import { ApolloServerPluginLandingPageLocalDefault } from "apollo-server-core";
 
 const users = require("../database/mockdb").users();
 const books = require("../database/mockdb").books();
 
-// A schema is a collection of type definitions (hence "typeDefs")
-// that together define the "shape" of queries that are executed against
-// your data.
-const typeDefs = gql`
-  # Comments in GraphQL strings (such as this one) start with the hash (#) symbol.
+const typeDefs = require("./graphq/schema").typeDefs();
 
-  # This "Book" type defines the queryable fields for every book in our data source.
-  type Book {
-    title: String
-    author: String
-  }
-
-  type User {
-    name: String
-    email: String
-    projects: [Project]
-  }
-
-  type Project {
-    title: String
-    active: Boolean!
-    members: [User]
-  }
-
-  # The "Query" type is special: it lists all of the available queries that
-  # clients can execute, along with the return type for each. In this
-  # case, the "books" query returns an array of zero or more Books (defined above).
-  type Query {
-    books: [Book]
-    users: [User]
-  }
-`;
 
 // Resolvers define how to fetch the types defined in your schema.
 // This resolver retrieves books from the "books" array above.
